Move all toolbar logic to app, where toolbar is mounted
